[PS3] Merge in Mokuno-san's gelic work.

+++ include/asm-powerpc/ps3.h
@@ -36,7 +36,7 @@ union ps3_firmware_version {
 };
 
 void ps3_get_firmware_version(union ps3_firmware_version *v);
-int ps3_compare_firmware_version(u16 major, u16 minor, u16 rev);
+s64 ps3_compare_firmware_version(u16 major, u16 minor, u16 rev);
 
 /* 'Other OS' area */
